+++ index.html
@@ -7,8 +7,6 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
-  <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
   <title>Кекстаграм</title>
 </head>
 
@@ -252,6 +250,7 @@
     </section>
   </template>
 
-</body>
-
+    <script src="/js/functions.js"></script>
+    <script src="/js/main.js" type="module"></script>
+  </body>
 </html>
