+++ js/const.js
@@ -1,0 +1,40 @@
+const DESCRIPTIONS = [
+  'Lake in forest', 'Road to the beach', 'Food',
+  'Mclaren', 'Plane', 'Shouse',
+  'Cat', 'Palm trees', 'Sea',
+  'Concert', 'Range-rover', 'Audi',
+  'Light', 'Juice', 'Girl',
+  'Ocean', 'Road', 'Sky',
+  'Sofa', 'Plym', 'plymouth',
+  'Red fish', 'Origin', 'Relax',
+  'Programming',
+];
+
+const NAMES = [
+  'Ivan', 'Oleg', 'Igor',
+  'Masha', 'Sasha', 'Pasha',
+  'Roma', 'Sergey', 'Olga',
+  'Irina', 'Misha', 'Polina',
+  'Evgenyi', 'Ekaterina', 'Milana',
+  'Adelina', 'Yra', 'Vasya', 'Sofa',
+  'Svetlana', 'Tatyana', 'Veronika',
+  'Valerya', 'Varvara', 'Vlad',
+];
+
+const COMMENTS = [
+  'Всё отлично!', 'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.', ' В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.', 'Как можно было поймать такой неудачный момент?!',
+];
+
+const MIN_VALUE = 1;
+const MAX_VALUE = 25;
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MAX_COMMENTS_VALUE = 30;
+const MAX_VALUE_AVATAR = 6;
+
+export {DESCRIPTIONS, NAMES, COMMENTS};
+export {MIN_VALUE, MAX_VALUE, MIN_LIKES, MAX_LIKES, MAX_COMMENTS_VALUE, MAX_VALUE_AVATAR};
